Блокировка фильтров в случае неуспешной загрузки данных похожих объявлений
