+++ js/api.js
@@ -1,6 +1,13 @@
 import {showAlert} from './util.js';
+import {enableFilters} from './filters.js';
 
-const getData = (onSuccess) => {fetch('https://23.javascript.pages.academy/keksobooking/dat')
+
+/**
+ * Получает данные похожих объявлений с сервера
+ *
+ * @param {function} onSuccess - колбэк, вызываемый в случае успешного выполнения запроса
+ */
+const getData = (onSuccess) => {fetch('https://23.javascript.pages.academy/keksobooking/data')
   .then((response) => {
     if (response.ok) {
       return response.json();
@@ -10,6 +17,7 @@ const getData = (onSuccess) => {fetch('https://23.javascript.pages.academy/kekso
   })
   .then((announcements) => {
     onSuccess(announcements);
+    enableFilters();
   });
 };
