+++ js/filters.js
@@ -1,0 +1,23 @@
+const mapFilters = document.querySelector('.map__filters');
+const mapFiltersSelect = mapFilters.querySelectorAll('select');
+const mapFiltersFieldset = mapFilters.querySelectorAll('fieldset');
+
+/** Переводит фильтры в неактивное состояние */
+const disableFilters = () => {
+  mapFilters.classList.add('map__filters--disabled');
+  mapFiltersSelect.forEach((element) => {element.setAttribute('disabled', 'disabled');
+  });
+  mapFiltersFieldset.forEach((element) => {element.setAttribute('disabled', 'disabled');
+  });
+};
+
+/** Переводит фильтры в активное состояние */
+const enableFilters = () => {
+  mapFilters.classList.remove('map__filters--disabled');
+  mapFiltersSelect.forEach((element) => {element.removeAttribute('disabled', 'disabled');
+  });
+  mapFiltersFieldset.forEach((element) => {element.removeAttribute('disabled', 'disabled');
+  });
+};
+
+export {disableFilters, enableFilters};
